LayoutTests/platform/glib/imported/w3c/web-platform-tests/svg/text/scripted/lengthadjust-expected.txt

Stretched text
Stretched text

FAIL Tests dynamic updates of the "lengthAdjust" content attribute assert_greater_than: expected a number greater than 86 but got 86
FAIL Tests dynamic updates of the "lengthAdjust" IDL attribute assert_greater_than: expected a number greater than 86 but got 86
